Make audio sample rate a double for accuracy
Also delete frameSampleCount entirely as it wasn't really good
for anything.

+++ FCEU/FCEUGameCore.mm
@@ -347,7 +347,7 @@ static void writeSaveFile(const char* path, int type)
     return GL_RGB5;
 }
 
-- (NSUInteger)frameSampleRate
+- (double)frameSampleRate
 {
     return SAMPLERATE;
 }
