+++ VisualBoyAdvance/GBAGameCore.mm
@@ -337,7 +337,7 @@ static void writeSaveFile(const char* path, int type)
     return GL_RGB5;
 }
 
-- (NSUInteger)frameSampleRate
+- (double)frameSampleRate
 {
     return SAMPLERATE;
 }
